requirements.txt
azure-ai-textanalytics==5.0.0
azure-cognitiveservices-language-textanalytics==0.2.0
azure-common==1.1.26
azure-core==1.9.0
azureml==0.2.7
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msrest==0.6.19
numpy==1.19.4
oauthlib==3.1.0
pandas==1.2.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
yapf==0.30.0
